fix: split policies for both legacy and cdk based REST APIs
